// shared/tdorig/tdorig-types.ts
export interface ITDoRig {
    codmonedamtodoc:string;
    iddoc: number;
    descdoc: string;
    numbenef: number;
    numidbenef: string;
    nombre: string;
    appabrev: string;
    tipodoc: string;
    desctipodoc: string;
    indreverso: string;
    stsdoc: string;
    descstsdoc: string;
    refdoc: string;
    fecdoc: string;
    ano: number;
    numop: string | null;
    iddocfis: string | null;
    mtodoc: string;
    codundadmpro: string;
    desccodundpro: string;
    descdocext: string;
  }